+++ NEWS.md
@@ -4,6 +4,9 @@
 
 - legend in `simple_dashboard()`.
 
+- new function `simple_dashboard2()`: same as `simple_dashboard()` but with smaller text to add informations on last outliers.
+
+
 # rjdqa 0.1.2
 
 - documentation updated.